// src/enums/option-type.ts
export enum OptionType {
    QueryPath = 'QueryPath',
    HasList = 'HasList',
    HasProfile = 'HasProfile',
    HasAdd = 'HasAdd',
    ServiceApi = 'ServiceApi',
    ListMode = 'ListMode',
    PreviewMode = 'PreviewMode',
    Plural = 'Plural',
    HasForm = 'HasForm',
    Property = 'Property',
    Flat = 'Flat',
    Model = 'Model',
    IndexModel = 'IndexModel',
    HasAttachment = 'HasAttachment',
    Name = 'Name',
    HasFilter = 'HasFilter',
    QueryServiceApi = 'QueryServiceApi',
}